requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==5.0.2
django-redis==5.4.0
djangorestframework==3.14.0
gunicorn==21.2.0
packaging==23.2
psycopg2-binary==2.9.9
pytz==2024.1
redis==5.0.1
sqlparse==0.4.4
